Migrate auth pages to server-side authentication
- Convert favorites, mypage, restaurants/new, and top pages to server components
- Remove client-side ProtectedPage wrapper in favor of getServerSession
- Extract TopPageClient component for client-side interactivity
- Improve error handling and loading states
